fix: moved shortucts only initialized once
Fix issuer with initializing shortcuts
for component delegation so the
shortcuts are not added with already
initialized actions.

Fixes #12307
